backend/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
facenet-pytorch==2.5.3
filelock==3.13.3
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
Jinja2==3.1.3
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mysqlclient==2.2.4
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
psycopg2==2.9.9